Refactor imports, move EVM method types to types
